Rendering overall
- resource sync from the engine AND the game to the build now sync readonly resources. This is to avoid confusing resources configs being erased after compilation
- moved setViewPort away from any particle 2d or 3d renders
- Framework (mainly opengl) buffers are now in the folder buffers
- material component will now reference a material. This is to create a Nx1 relation between objects and materials instead of 1x1. Meaning material can be reused to avoid exploding the program
- create instancing with cube meshes
- fix minor bugs inside the spray batch
- shaders uniforms can now be set to ignore any errors
- add mesh component, it can now setup cubes and spheres
- change opengl backend to support instanced elements
- smooth transition between 2d and 3d was tried, but this need to be redone. This commit disables 2d rendering for now
- fix framebuffer not rendering 3d meshes
- fix vertex attrib pointer not allowing 3x3 and 4x4 matrices into the layout
- Material manager created to handle and owning shaders and materials

+++ Pain/src/CoreRender/Renderer/BatchLight.cpp
@@ -40,8 +40,8 @@ void PointLightBatch::upload(Shader &shader)
 {
   shader.bind();
 
-  uint32_t count = static_cast<uint32_t>(m_positions.size());
-  shader.uploadUniformInt("u_PointLightCount", count);
+  unsigned long count = m_positions.size();
+  shader.uploadUniformInt("u_PointLightCount", static_cast<int32_t>(count));
 
   for (uint32_t i = 0; i < count; i++) {
     shader.uploadUniformFloat4(
